requirements.txt
# django
Django==3.0.3
django-environ==0.4.5

# google
google-cloud==0.34.0
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-bigquery==2.32.0
google-cloud-core==2.2.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0

# ipython
ipython==8.0.1

# postgreSQL
psycopg2==2.8.4
psycopg2-binary==2.8.6

# restframework
djangorestframework==3.12.2

# swagger
drf-yasg==1.20.0

# uwsgi
uwsgi==2.0.20